+++ doc/version_history.rst
@@ -2,6 +2,17 @@
 Version History
 ===============
 
+v0.13.1
+-------
+
+* Bypass the actions if the GUI is connecting to the controller.
+* Fix the bug of setting the hardpoints and configuration file. 
+
+v0.13.0
+-------
+
+* Read the following settings from the **ts_config_mttcs**: inclination source, configuration file, and hardpoints.
+
 v0.12.6
 -------